Refactor to render same form component for new && edit
