+++ app/templates/reminders/edit.hbs
@@ -1,1 +1,2 @@
-{{edit-reminder}}
+<h3> Edit Reminder Below </h3>
+{{reminder-form}}
