+++ app/templates/reminders/new.hbs
@@ -1,1 +1,2 @@
-{{new-reminder-form}}
+
+{{reminder-form}}
